{"schema":"decodex/commit/1","summary":"Collapse over-split MCP and review helper modules","authority":"manual"}
